small ui changes that were requested
