+++ packages/admin-frontend/components/Container.js
@@ -5,8 +5,6 @@ const Container = (props: {}) => (
   <div
     css={{
       maxWidth: '960px',
-      margin: '0 auto',
-      paddingTop: 80,
       paddingLeft: 16,
       paddingRight: 16
     }}
