+++ packages/admin-frontend/style/media.js
@@ -1,4 +1,3 @@
-// @flow
 /**
  * Part of GDL gdl-frontend.
  * Copyright (C) 2017 GDL
